Refactor LLM instructions: append model with ccli config llm-instructions
Breaking change to system.llm.instructions behavior:
- Custom value is now APPENDED as a PROJECT CONTEXT section, not a
  replacement of the entire default instructions
- Built-in defaults (schema, scope, tool tips, effective usage) stay
  immutable and up to date as features are added

New: ccli config llm-instructions
- Shows effective instructions (defaults + custom additions)
- --default flag shows just the built-in defaults

Extracted shared module src/llm-instructions.ts so both MCP server
and CLI can access the same instruction logic.

501 tests passing.
